presentation/src/test/java/com/lawmobile/presentation/utils/CameraHelperTest.kt
package com.lawmobile.presentation.utils

import com.safefleet.mobile.avml.cameras.external.CameraConnectService
import io.mockk.clearAllMocks
import io.mockk.every
import io.mockk.mockk
import org.junit.Assert
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestInstance

@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class CameraHelperTest {

    private val wifiHelper: WifiHelper = mockk()
    private val cameraConnectService: CameraConnectService = mockk()
    private val cameraHelper = CameraHelper(cameraConnectService, wifiHelper)
    
    @BeforeEach
    fun setUp(){
        clearAllMocks()
    }

    @Test
    fun testSetInstance() {
        CameraHelper.setInstance(cameraHelper)
        Assert.assertEquals(CameraHelper.getInstance(), cameraHelper)
    }

    @Test
    fun testCheckWithAlertIfTheCameraIsConnectedSuccess() {
        every { wifiHelper.getGatewayAddress() } returns ""
        every { cameraConnectService.isCameraConnected(any()) } returns true
        Assert.assertTrue(cameraHelper.checkWithAlertIfTheCameraIsConnected())
    }

    @Test
    fun testCheckWithAlertIfTheCameraIsConnectedFailed() {
        every { wifiHelper.getGatewayAddress() } returns ""
        every { cameraConnectService.isCameraConnected(any()) } returns false
        Assert.assertFalse(cameraHelper.checkWithAlertIfTheCameraIsConnected())
    }
}